+++ examples/example/vue3/src/examples/dropdown/01/App.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { ref } from 'vue'
-import { Button } from '@expcat/tigercat-vue/Button'
 import { Dropdown } from '@expcat/tigercat-vue/Dropdown'
 import { DropdownItem } from '@expcat/tigercat-vue/DropdownItem'
 import { DropdownMenu } from '@expcat/tigercat-vue/DropdownMenu'
@@ -10,11 +9,12 @@ const lastAction = ref('尚未选择')
 
 <template>
   <div class="space-y-3">
-    <Dropdown trigger="click">
-      <Button>操作</Button>
+    <Dropdown>
+      操作
       <DropdownMenu>
         <DropdownItem @click="lastAction = '编辑'">编辑</DropdownItem>
         <DropdownItem @click="lastAction = '复制'">复制</DropdownItem>
+        <DropdownItem disabled>不可用</DropdownItem>
         <DropdownItem divided @click="lastAction = '删除'">删除</DropdownItem>
       </DropdownMenu>
     </Dropdown>